Switch to 64-bits everywhere

+++ natural/factors_test.go
@@ -5,8 +5,8 @@ import (
 	"testing"
 )
 
-func allFactors(x int64) []int32 {
-	fact := make([]int32, 0, 8)
+func allFactors(x int64) []int64 {
+	fact := make([]int64, 0, 8)
 	for f := range Factors(x) {
 		fact = append(fact, f)
 	}
@@ -16,19 +16,19 @@ func allFactors(x int64) []int32 {
 func TestFactors(t *testing.T) {
 
 	fact := allFactors(13195)
-	exp := []int32{5, 7, 13, 29}
+	exp := []int64{5, 7, 13, 29}
 	if !reflect.DeepEqual(fact, exp) {
 		t.Error("invalid factors", fact)
 	}
 
 	fact = allFactors(14)
-	exp = []int32{2, 7}
+	exp = []int64{2, 7}
 	if !reflect.DeepEqual(fact, exp) {
 		t.Error("invalid factors", fact)
 	}
 
 	fact = allFactors(48)
-	exp = []int32{2, 2, 2, 2, 3}
+	exp = []int64{2, 2, 2, 2, 3}
 	if !reflect.DeepEqual(fact, exp) {
 		t.Error("invalid factors", fact)
 	}
