+++ p40_test.go
@@ -8,7 +8,7 @@ import (
 func TestP40(t *testing.T) {
 
 	dig := make([]byte, 1, 1100000)
-	for i := 1; len(dig) <= 1000000; i++ {
+	for i := int64(1); len(dig) <= 1000000; i++ {
 		idig := natural.Digits(i)
 		natural.Reverse(idig)
 		dig = append(dig, idig...)
